# Low-hanging fruit performance improvements of the hubble consumer module