--- a/pkg/hubble/monitor/consumer.go
+++ b/pkg/hubble/monitor/consumer.go
@@ -54,14 +54,15 @@ func NewConsumer(observer Observer) monitorConsumer.MonitorConsumer {
 
 // sendEventQueueLostEvents tries to send the current value of the lost events
 // counter to the observer. If it succeeds to enqueue a notification, it
-// resets the counter.
-func (c *consumer) sendNumLostEvents() {
+// resets the counter. Returns a boolean indicating whether the notification
+// has been successfully sent.
+func (c *consumer) sendNumLostEvents() bool {
 	c.lostLock.Lock()
 	defer c.lostLock.Unlock()
 	// check again, in case multiple
 	// routines contended the lock
 	if c.numEventsLost == 0 {
-		return
+		return true
 	}
 
 	numEventsLostNotification := &observerTypes.MonitorEvent{
@@ -79,24 +80,42 @@ func (c *consumer) sendNumLostEvents() {
 		// successfully notified the observer about the amount of events
 		// that were lost since the previous LostEvent message
 		c.numEventsLost = 0
+		return true
 	default:
 		// We do not need to bump the numEventsLost counter here, as we will
 		// try to send a new LostEvent notification again during the next
 		// invocation of sendEvent
+		return false
 	}
 }
 
 // sendEvent enqueues an event in the observer. If this is not possible, it
 // keeps a counter of lost events, which it will regularly try to send to the
 // observer as well
-func (c *consumer) sendEvent(event *observerTypes.MonitorEvent) {
+func (c *consumer) sendEvent(payloader func() interface{}) {
 	if c.numEventsLost > 0 {
-		c.sendNumLostEvents()
+		if !c.sendNumLostEvents() {
+			// We just failed sending the lost notification, hence it doesn't
+			// make sense to try and send the actual event, as we'll most
+			// likely fail as well.
+			c.countDroppedEvent()
+			return
+		}
 	}
+
 	select {
-	case c.observer.GetEventsChannel() <- event:
+	case c.observer.GetEventsChannel() <- c.newEvent(payloader):
 	default:
 		c.countDroppedEvent()
+	}
+}
+
+func (c *consumer) newEvent(payloader func() interface{}) *observerTypes.MonitorEvent {
+	return &observerTypes.MonitorEvent{
+		UUID:      uuid.New(),
+		Timestamp: time.Now(),
+		NodeName:  nodeTypes.GetAbsoluteNodeName(),
+		Payload:   payloader(),
 	}
 }
 
@@ -115,41 +134,32 @@ func (c *consumer) countDroppedEvent() {
 
 // NotifyAgentEvent implements monitorConsumer.MonitorConsumer
 func (c *consumer) NotifyAgentEvent(typ int, message interface{}) {
-	c.sendEvent(&observerTypes.MonitorEvent{
-		UUID:      uuid.New(),
-		Timestamp: time.Now(),
-		NodeName:  nodeTypes.GetAbsoluteNodeName(),
-		Payload: &observerTypes.AgentEvent{
+	c.sendEvent(func() interface{} {
+		return &observerTypes.AgentEvent{
 			Type:    typ,
 			Message: message,
-		},
+		}
 	})
 }
 
 // NotifyPerfEvent implements monitorConsumer.MonitorConsumer
 func (c *consumer) NotifyPerfEvent(data []byte, cpu int) {
-	c.sendEvent(&observerTypes.MonitorEvent{
-		UUID:      uuid.New(),
-		Timestamp: time.Now(),
-		NodeName:  nodeTypes.GetAbsoluteNodeName(),
-		Payload: &observerTypes.PerfEvent{
+	c.sendEvent(func() interface{} {
+		return &observerTypes.PerfEvent{
 			Data: data,
 			CPU:  cpu,
-		},
+		}
 	})
 }
 
 // NotifyPerfEventLost implements monitorConsumer.MonitorConsumer
 func (c *consumer) NotifyPerfEventLost(numLostEvents uint64, cpu int) {
-	c.sendEvent(&observerTypes.MonitorEvent{
-		UUID:      uuid.New(),
-		Timestamp: time.Now(),
-		NodeName:  nodeTypes.GetAbsoluteNodeName(),
-		Payload: &observerTypes.LostEvent{
+	c.sendEvent(func() interface{} {
+		return &observerTypes.LostEvent{
 			Source:        observerTypes.LostEventSourcePerfRingBuffer,
 			NumLostEvents: numLostEvents,
 			CPU:           cpu,
-		},
+		}
 	})
 	c.metricLostPerfEvents.Inc()
 }