See individual commits for additional details.

```
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/hubble/monitor
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
                                     │   old.txt    │               new.txt               │
                                     │    sec/op    │   sec/op     vs base                │
HubbleConsumerSendEvent/all_sent-20     426.5n ± 6%   156.6n ± 2%  -63.28% (p=0.000 n=10)
HubbleConsumerSendEvent/all_lost-20    968.05n ± 3%   57.59n ± 2%  -94.05% (p=0.000 n=10)
HubbleConsumerSendEvent/half_sent-20   1026.0n ± 1%   208.2n ± 2%  -79.71% (p=0.000 n=10)
geomean                                 751.0n        123.4n       -83.58%

                                     │  old.txt   │                 new.txt                 │
                                     │    B/op    │    B/op     vs base                     │
HubbleConsumerSendEvent/all_sent-20    128.0 ± 0%   112.0 ± 0%   -12.50% (p=0.000 n=10)
HubbleConsumerSendEvent/all_lost-20    272.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
HubbleConsumerSendEvent/half_sent-20   272.0 ± 0%   108.0 ± 0%   -60.29% (p=0.000 n=10)
geomean                                211.6                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                                     │  old.txt   │                 new.txt                 │
                                     │ allocs/op  │ allocs/op   vs base                     │
HubbleConsumerSendEvent/all_sent-20    3.000 ± 0%   2.000 ± 0%   -33.33% (p=0.000 n=10)
HubbleConsumerSendEvent/all_lost-20    7.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
HubbleConsumerSendEvent/half_sent-20   7.000 ± 0%   2.000 ± 0%   -71.43% (p=0.000 n=10)
geomean                                5.278                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```