@rolinh 2024-09-06
> @doniacld Gentle ping 🙏

David already approved for sig agent, I think we're good.